LayoutTests/inspector/debugger/watch-expressions-preserve-expansion-expected.txt

Test that watch expressions expansion state is restored after update.

Bug 99304
Debugger was enabled.
Watch expressions added.
expanded globalObject [object Object]
expanded foo [object Object]
expanded bar [object Object]
Watch expressions expanded.
  globalObject: Object
    foo: Object
      bar: Object
        baz: 2012
        __proto__: Object
      __proto__: Object
    __proto__: Object
  windowAlias: Window
error: Execution context with given id not found.
error: Execution context with given id not found.
Page reloaded.
Watch expressions after page reload:
  globalObject: Object
    foo: Object
      bar: Object
        baz: 2012
        __proto__: Object
      __proto__: Object
    __proto__: Object
  windowAlias: Window
Debugger was disabled.
